anhanga/requirements.txt
flask==3.1.3
flask-sqlalchemy==3.1.1
sqlalchemy==3.1.1
fastapi==0.136.0
uvicorn==0.44.0
pydantic==2.13.1
requests==2.32.3
python-dotenv==1.0.1
